+++ packages/article-headline/article-headline.stories.js
@@ -5,5 +5,5 @@ import { storiesOf } from "@storybook/react-native";
 import ArticleHeadline from "./article-headline";
 
 storiesOf("ArticleHeadline", module).add("ArticleHeadline", () => (
-  <ArticleHeadline title="Labour MPs urge Jeremy Corbyn to condemn Maduro’s Venezuela regime" />
+  <ArticleHeadline text="Labour MPs urge Jeremy Corbyn to condemn Maduro’s Venezuela regime" />
 ));