+++ packages/article/article.js
@@ -1,29 +1,18 @@
 import React from "react";
-import {
-  Platform,
-  Text,
-  View,
-  ListView,
-  ActivityIndicator
-} from "react-native";
+import { Platform, Text, View, ListView } from "react-native";
 import PropTypes from "prop-types";
 import { renderTrees } from "@times-components/markup";
 import Image from "@times-components/image";
 import ArticleImage from "@times-components/article-image";
 import { AdComposer } from "@times-components/ad";
 
+import ArticleError from "./article-error";
+import ArticleLoading from "./article-loading";
+
 import listViewDataHelper from "./data-helper";
 import styles from "./styles/article-style";
 import ArticleHeader from "./article-header";
 import ArticleMeta from "./article-meta";
-import {
-  articleHeaderPropTypes,
-  articleHeaderDefaultPropTypes
-} from "./article-header.proptypes";
-import {
-  articleMetaPropTypes,
-  articleMetaDefaultPropTypes
-} from "./article-meta.proptypes";
 
 const ds = new ListView.DataSource({ rowHasChanged: (r1, r2) => r1 !== r2 });
 const listViewPageSize = 1;
@@ -43,10 +32,10 @@ class ArticlePage extends React.Component {
         </View>
       );
     } else if (rowData.type === "header") {
-      const { title, flags, standfirst, label } = rowData.data;
+      const { headline, flags, standfirst, label } = rowData.data;
       return (
         <ArticleHeader
-          title={title}
+          headline={headline}
           flags={flags}
           standfirst={standfirst}
           label={label}
@@ -99,12 +88,13 @@ class ArticlePage extends React.Component {
 
     return null;
   }
+
   constructor(props) {
     super(props);
 
-    if (props.data.article) {
+    if (props.article && !props.isLoading && !props.error) {
       this.state = {
-        dataSource: ds.cloneWithRows(listViewDataHelper(props.data.article))
+        dataSource: ds.cloneWithRows(listViewDataHelper(props.article))
       };
     } else {
       this.state = {
@@ -114,20 +104,22 @@ class ArticlePage extends React.Component {
   }
 
   componentWillReceiveProps(nextProps) {
-    if (!nextProps.isLoading && nextProps.data) {
+    if (!nextProps.isLoading && !nextProps.error) {
       this.setState({
-        dataSource: ds.cloneWithRows(listViewDataHelper(nextProps.data.article))
+        dataSource: ds.cloneWithRows(listViewDataHelper(nextProps.article))
       });
     }
   }
 
   render() {
-    if (this.props.isLoading) {
-      return (
-        <View style={styles.container}>
-          <ActivityIndicator size={"large"} />
-        </View>
-      );
+    const { error, isLoading } = this.props;
+
+    if (error) {
+      return <ArticleError {...error} />;
+    }
+
+    if (isLoading) {
+      return <ArticleLoading />;
     }
 
     const ArticleListView = (
@@ -149,23 +141,12 @@ class ArticlePage extends React.Component {
 }
 
 ArticlePage.propTypes = {
-  data: PropTypes.shape({
-    article: PropTypes.shape({
-      ...articleHeaderPropTypes,
-      ...articleMetaPropTypes
-    })
+  article: PropTypes.shape({
+    ...ArticleHeader.propTypes,
+    ...ArticleMeta.propTypes
   }),
-  isLoading: PropTypes.bool
-};
-
-ArticlePage.defaultProps = {
-  data: {
-    article: {
-      ...articleHeaderDefaultPropTypes,
-      ...articleMetaDefaultPropTypes
-    }
-  },
-  isLoading: true
+  isLoading: PropTypes.bool,
+  error: PropTypes.shape()
 };
 
 export default ArticlePage;