+++ packages/card/card.stories.js
@@ -1,23 +1,20 @@
-import { View } from "react-native";
 import React from "react";
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { storiesOf } from "@storybook/react-native";
 import Card from "./card";
 import props from "./fixtures/card-props.json";
-
-const story = m => <View style={{ padding: 20 }}>{m}</View>;
+import LateralSpacingDecorator from "../../storybook/decorators/lateral-spacing";
 
 props.date = new Date("2017-07-01T14:32:00.000Z");
 storiesOf("Card", module)
-  .add("Card", () => story(<Card {...props} />))
-  .add("Card without Image", () =>
-    story(
-      <Card
-        {...Object.assign({}, props, {
-          image: {
-            url: null
-          }
-        })}
-      />
-    )
-  );
+  .addDecorator(LateralSpacingDecorator)
+  .add("Card", () => <Card {...props} />)
+  .add("Card without Image", () => (
+    <Card
+      {...Object.assign({}, props, {
+        image: {
+          url: null
+        }
+      })}
+    />
+  ));