+++ packages/article-headline/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.8.0"></a>
+# [0.8.0](https://github.com/newsuk/times-components/compare/@times-components/article-headline@0.7.2...@times-components/article-headline@0.8.0) (2017-10-16)
+
+
+### Features
+
+* article with provider ([#283](https://github.com/newsuk/times-components/issues/283)) ([d0c0792](https://github.com/newsuk/times-components/commit/d0c0792))
+
+
+
+
+<a name="0.7.2"></a>
+## [0.7.2](https://github.com/newsuk/times-components/compare/@times-components/article-headline@0.7.1...@times-components/article-headline@0.7.2) (2017-10-13)
+
+
+### Bug Fixes
+
+* prettier to ignore native code ([#296](https://github.com/newsuk/times-components/issues/296)) ([d6d24ed](https://github.com/newsuk/times-components/commit/d6d24ed))
+
+
+
+
 <a name="0.7.1"></a>
 ## [0.7.1](https://github.com/newsuk/times-components/compare/@times-components/article-headline@0.7.0...@times-components/article-headline@0.7.1) (2017-10-13)