+++ packages/tracking/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.2.0"></a>
+# [0.2.0](https://github.com/newsuk/times-components/compare/@times-components/tracking@0.1.2...@times-components/tracking@0.2.0) (2017-10-13)
+
+
+### Features
+
+* track events ([#276](https://github.com/newsuk/times-components/issues/276)) ([df67de2](https://github.com/newsuk/times-components/commit/df67de2))
+
+
+
+
+<a name="0.1.2"></a>
+## [0.1.2](https://github.com/newsuk/times-components/compare/@times-components/tracking@0.1.1...@times-components/tracking@0.1.2) (2017-10-13)
+
+
+### Bug Fixes
+
+* prettier to ignore native code ([#296](https://github.com/newsuk/times-components/issues/296)) ([d6d24ed](https://github.com/newsuk/times-components/commit/d6d24ed))
+
+
+
+
 <a name="0.1.1"></a>
 ## [0.1.1](https://github.com/newsuk/times-components/compare/@times-components/tracking@0.1.0...@times-components/tracking@0.1.1) (2017-10-13)